fix(ci): clear go-lint, go-test race, and swift-lint failures on 0.5.1
- gofmt struct-literal alignment in briefing/db tests
- rename `real` → `hasItems`/`hasTracks` (predeclared identifier shadow)
- defer rows.Close() in NotifyDueTargets (sqlclosecheck)
- drop redundant int64() conversion in inbox_more_test (unconvert)
- guard codex.resetCache() with a mutex so TestResetCache_RaceFree is race-free
- swiftlint disable force_try in GuideDetailViewTests.makeGuide JSON helper
- convert trailing-closure call sites in GuideDetailViewTests / WorkloadPersonDetailViewTests

+++ internal/db/inbox_more_test.go
@@ -67,7 +67,7 @@ func TestGetInboxItem_Int64(t *testing.T) {
 
 	got, err := db.GetInboxItem(id)
 	require.NoError(t, err)
-	assert.Equal(t, int64(id), int64(got.ID))
+	assert.Equal(t, id, int64(got.ID))
 	assert.Equal(t, "needs reply", got.Snippet)
 }
